refactor: home view 분리
