+++ package.json
@@ -16,12 +16,13 @@
     "axios": "^1.4.0",
     "lodash": "^4.17.21",
     "next": "13.5.4",
+    "pnpm": "^8.12.0",
     "react": "^18.2.0",
     "react-cookie": "^6.1.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.43.4",
     "react-query": "^3.39.3",
-    "styled-components": "^5.3.6",
+    "styled-components": "^5.3.11",
     "yup": "^1.3.2"
   },
   "devDependencies": {
