+++ src/app/layout.tsx
@@ -1,27 +1,22 @@
-import type { Metadata } from 'next';
-
-import Provider from '@/components/common/Provider';
-import ThemeProvider from '@/components/common/ThemeProvider';
-
 import ErrorBoundary from '@/components/common/ErrorBoundary';
+import QueryProvider from '@/components/common/QueryProvider';
+import ThemeProvider from '@/components/common/ThemeProvider';
+import type { Metadata } from 'next';
+import { PropsWithChildren } from 'react';
 
 export const metadata: Metadata = {
   title: 'Daehwa',
   description: 'Daehwa app',
 };
 
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
+export default function RootLayout({ children }: PropsWithChildren) {
   return (
     <html lang="ko">
       <body>
         <ErrorBoundary>
-          <Provider>
+          <QueryProvider>
             <ThemeProvider>{children}</ThemeProvider>
-          </Provider>
+          </QueryProvider>
         </ErrorBoundary>
       </body>
     </html>
