+++ src/components/common/Fallback.tsx
@@ -4,8 +4,6 @@ interface Props {
   error?: Error | null;
 }
 
-const Fallback = ({ error }: Props) => {
+export const Fallback = ({ error }: Props) => {
   return <div>에러: {error?.message}</div>;
 };
-
-export default Fallback;
