+++ src/components/common/index.ts
@@ -1,7 +1,5 @@
 export * from './AuthButton';
 export * from './AuthInput';
 export * from './Balloon';
-export * from './ErrorBoundary';
 export * from './Fallback';
-export * from './ThemeProvider';
 export * from './ClientComponentLayout';
