+++ src/styled.d.ts
@@ -1,8 +1,8 @@
 import 'styled-components';
 import theme from '@/styles/theme';
 
-type CustomTheme = typeof theme;
+type Theme = typeof theme;
 
 declare module 'styled-components' {
-  export interface DefaultTheme extends CustomTheme {}
+  export interface DefaultTheme extends Theme {}
 }
