+++ src/styles/globalStyle.ts
@@ -2,8 +2,6 @@ import { createGlobalStyle } from 'styled-components';
 import theme from './theme';
 
 const GlobalStyle = createGlobalStyle`
- @import url('https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.6/dist/web/static/pretendard-dynamic-subset.css');
-
   * {
     box-sizing: border-box;
   }
