+++ src/styles/theme.ts
@@ -4,6 +4,6 @@ import { spacing } from './spacing';
 const theme = {
   colors,
   spacing,
-};
+} as const;
 
 export default theme;
